Conflicts:
	WebContent/WEB-INF/lib/semi.sql
	WebContent/rent/return_ok.jsp
	src/kosta/albatross/book/controllers/BookRegisterController.java
	src/kosta/albatross/member/controllers/MemberUpdateController.java
	src/kosta/albatross/member/controllers/RegisterController.java
	src/kosta/albatross/member/controllers/memberFindViewController.java
	src/kosta/albatross/member/controllers/memberId_pwCheckServlet.java
	src/kosta/albatross/rent/controllers/ReturnBookController.java
	src/kosta/albatross/rent/models/RentDAO.java

+++ src/kosta/albatross/book/models/BookDAO.java
@@ -10,6 +10,8 @@ import javax.sql.DataSource;
 
 import kosta.albatross.common.models.DataSourceManager;
 import kosta.albatross.common.models.PagingBean;
+import kosta.albatross.exception.NotDeleteBookException;
+import kosta.albatross.rent.models.RentDAO;
 
 public class BookDAO {
 
@@ -376,11 +378,15 @@ public class BookDAO {
 	 * 
 	 * @param bNo
 	 * @throws SQLException
+	 * @throws NotDeleteBookException 
 	 */
-	public void deleteBook(int bNo) throws SQLException {
+	public void deleteBook(int bNo) throws SQLException, NotDeleteBookException {
 		Connection con = null;
 		PreparedStatement pstmt = null;
 		try {
+			if(RentDAO.getInstance().isRentingBook(bNo)) {
+				throw new NotDeleteBookException();
+			}
 			con = dataSource.getConnection();
 			String sql = "DELETE FROM SEMI_BOOK WHERE bNo = ?";
 			pstmt = con.prepareStatement(sql);
@@ -420,7 +426,6 @@ public class BookDAO {
 			if (rs.next()) {
 				BVO = getBookDetail(rs.getInt(1));
 			}
-
 		} finally {
 			closeAll(rs, pstmt, con);
 		}
